pyproject.toml
[tool.black]
target-version = ['py310']
include = '\.pyi?$'
skip-string-normalization = true
exclude = '''
/(
    \.git
)/
'''

[tool.poetry]
name = "iotserver"
version = "0.9.0"
description = "Simple IoT Server, Configuration Tool & Dashboard"
authors = ["Jonathan Bydendyk <jpbydendyk@hacklab.co.za>"]
license = "BSD"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.4"
click = "^8.1.2"
django-map-widgets = "^0.3.3"
djangorestframework = "^3.13.1"
djangorestframework-gis = "^0.18"
paho-mqtt = "^1.6.1"
requests = "^2.27.1"
python-dotenv = "^0.20.0"
pymemcache = "^3.5.2"
psycopg2-binary = "^2.9.3"
django-cors-headers = "^3.12.0"
django-filter = "^22.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pytest-django = "^4.5.2"
black = "^22.3.0"
factory-boy = "^3.2.1"
pytest-mock = "^3.7.0"
pre-commit = "^2.18.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
